+++ src/main/java/raf/draft/dsw/controller/states/MoveState.java
@@ -7,7 +7,7 @@ import java.awt.geom.AffineTransform;
 
 public class MoveState extends AbstractState{
     @Override
-    public void mouseDragged(int dx, int dy, VisualElement element, RoomTab roomTab) {
+    public void mouseDragged(double dx, double dy, VisualElement element, RoomTab roomTab) {
         if (element != null) element.translate(dx, dy);
         else{
             roomTab.preConcatenateTransform(AffineTransform.getTranslateInstance(dx, dy));
